Move post request to folder POST
