+++ web/controller/controller.js
@@ -14,7 +14,7 @@ const registration_worker = require('../models/Api/registration_worker');
 const book_table = require('../models/Api/book_table');
 const add_storehouse = require('../models/Api/add_storehouse');
 const add_client = require('../models/Api/add_client');
-const add_client_order = require('../models/Api/add_client_order');
+const add_client_order = require('../models/Api/POST/add_client_order');
 const change_order_status = require('../models/Api/change_order_status');
 const record_giving_time = require('../models/Api/record_giving_time');
 const cancel_booking = require('../models/Api/cancel_booking');
